remove modelconfig from stratified-matrix
